GEODE-5787: extract ProcessHolder to be an outer class
  * Removed getter on ProcessHolder to get underlying  process
  * Added waitFor to ProcessHolder
  * Added getters to get underlying process error & input stream
