+++ geode-dunit/src/main/java/org/apache/geode/test/dunit/standalone/ProcessManager.java
@@ -95,8 +95,8 @@ public class ProcessManager {
       pendingVMs++;
       ProcessHolder holder = new ProcessHolder(process);
       processes.put(vmNum, holder);
-      linkStreams(version, vmNum, holder, process.getErrorStream(), System.err);
-      linkStreams(version, vmNum, holder, process.getInputStream(), System.out);
+      linkStreams(version, vmNum, holder, holder.getErrorStream(), System.err);
+      linkStreams(version, vmNum, holder, holder.getInputStream(), System.out);
     } catch (RuntimeException | Error t) {
       t.printStackTrace();
       throw t;
@@ -141,7 +141,7 @@ public class ProcessManager {
       } else {
         holder.kill();
       }
-      holder.getProcess().waitFor();
+      holder.waitFor();
       System.out.println("Old process for vm_" + vmNum + " has exited");
       launchVM(version, vmNum, true);
     } catch (InterruptedException | IOException e) {
@@ -326,37 +326,6 @@ public class ProcessManager {
     }
 
     return true;
-  }
-
-  public static class ProcessHolder {
-    private final Process process;
-    private volatile boolean killed = false;
-
-    public ProcessHolder(Process process) {
-      this.process = process;
-    }
-
-    public void kill() {
-      this.killed = true;
-      process.destroy();
-    }
-
-    public void killForcibly() {
-      this.killed = true;
-      process.destroyForcibly();
-    }
-
-    public Process getProcess() {
-      return process;
-    }
-
-    public boolean isKilled() {
-      return killed;
-    }
-
-    public boolean isAlive() {
-      return !killed && process.isAlive();
-    }
   }
 
   public RemoteDUnitVMIF getStub(int i)
